# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Knowledge OS Read-only API
  version: 1.0.0
  description: Web 与未来 App 共用的只读 HTTP API v1；所有知识查询只返回 public 数据。
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: 本机默认服务
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        '200':
          description: 服务与只读数据库状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /taxonomy:
    get:
      operationId: getTaxonomy
      responses:
        '200':
          description: 可公开分类树
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxonomyResponse'
  /documents:
    get:
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 分页知识摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPageResponse'
  /documents/{id}:
    get:
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: 知识详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /search:
    get:
      operationId: searchDocuments
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 分页搜索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPageResponse'
components:
  parameters:
    Query:
      name: query
      in: query
      required: false
      schema:
        type: string
        default: ''
        maxLength: 200
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    NotFound:
      description: 知识条目不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    HealthResponse:
      type: object
      required: [api_version, data]
      properties:
        api_version: { type: string, const: v1 }
        data: { $ref: '#/components/schemas/HealthStatus' }
    TaxonomyResponse:
      type: object
      required: [api_version, data]
      properties:
        api_version: { type: string, const: v1 }
        data:
          type: array
          items: { $ref: '#/components/schemas/TaxonomyNode' }
    DocumentPageResponse:
      type: object
      required: [api_version, data]
      properties:
        api_version: { type: string, const: v1 }
        data: { $ref: '#/components/schemas/DocumentPage' }
    DocumentResponse:
      type: object
      required: [api_version, data]
      properties:
        api_version: { type: string, const: v1 }
        data: { $ref: '#/components/schemas/KnowledgeDocument' }
    SearchPageResponse:
      type: object
      required: [api_version, data]
      properties:
        api_version: { type: string, const: v1 }
        data: { $ref: '#/components/schemas/SearchPage' }
    ApiError:
      type: object
      required: [api_version, error]
      properties:
        api_version: { type: string, const: v1 }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    HealthStatus:
      type: object
      required: [status, service, database, schema_version]
      properties:
        status: { type: string }
        service: { type: string }
        database: { type: string }
        schema_version: { type: integer }
    TaxonomyNode:
      type: object
      required: [id, parent_id, name, path, locked, children]
      properties:
        id: { type: string }
        parent_id: { type: [string, 'null'] }
        name: { type: string }
        path: { type: array, items: { type: string } }
        locked: { type: boolean }
        children: { type: array, items: { $ref: '#/components/schemas/TaxonomyNode' } }
    KnowledgeDocumentSummary:
      type: object
      required: [id, title, summary, tags, node_id, path, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        tags: { type: array, items: { type: string } }
        node_id: { type: string }
        path: { type: array, items: { type: string } }
        updated_at: { type: string }
    KnowledgeDocument:
      allOf:
        - $ref: '#/components/schemas/KnowledgeDocumentSummary'
        - type: object
          required: [body, source, relations]
          properties:
            body: { type: string }
            source: { $ref: '#/components/schemas/KnowledgeSource' }
            relations:
              type: array
              items: { $ref: '#/components/schemas/RelationEdge' }
    KnowledgeSearchHit:
      type: object
      required: [id, title, highlighted_title, summary, snippet, tags, node_id, path, rank, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        highlighted_title:
          type: string
          description: 匹配内容用纯文本 [[...]] 标记，不包含 HTML
        summary: { type: string }
        snippet:
          type: string
          description: 搜索摘要，匹配内容用纯文本 [[...]] 标记
        tags: { type: array, items: { type: string } }
        node_id: { type: string }
        path: { type: array, items: { type: string } }
        rank:
          type: number
          description: 相关度排序值，数值越小越相关
        updated_at: { type: string }
    KnowledgeSource:
      type: object
      required: [kind, original_name, sha256]
      properties:
        kind: { type: string }
        original_name: { type: string }
        sha256: { type: string }
    RelationEdge:
      type: object
      required: [id, from_node_id, to_node_id, relation_type, label, confidence]
      properties:
        id: { type: string }
        from_node_id: { type: string }
        to_node_id: { type: string }
        relation_type: { type: string }
        label: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
    DocumentPage:
      type: object
      required: [page, size, total, items]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        items:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeDocumentSummary' }
    SearchPage:
      type: object
      required: [page, size, total, items]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        items:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeSearchHit' }